requirements.txt
openai>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
fpdf2>=2.8.0
pytest>=8.0.0
